requirements.txt
pytest==5.2.1
google-cloud-storage==1.20.0
google-cloud-bigquery==1.20.0